refactor: modularize retrieval components
