+++ tests/test_question_metadata.py
@@ -3,7 +3,7 @@ from pathlib import Path
 
 sys.path.append(str(Path(__file__).resolve().parents[1]))
 
-from OcchioOnniveggente.src.retrieval import load_questions
+from src.retrieval import load_questions
 
 
 def test_question_metadata_fields():
