+++ tests/test_retrieval_basic.py
@@ -3,7 +3,7 @@ from pathlib import Path
 
 sys.path.append(str(Path(__file__).resolve().parents[1]))
 
-from OcchioOnniveggente.src.retrieval import retrieve
+from src.retrieval import retrieve
 
 
 class DummyStore:
